+++ versions/1.21.5/src/main/java/at/hannibal2/skyhanni/utils/render/RoundedShapeDrawer.kt
@@ -65,6 +65,18 @@ object RoundedShapeDrawer {
             postVertexOps = listOf { color(color) },
         )
 
+    fun drawRoundedRect(left: Int, top: Int, right: Int, bottom: Int, topColor: Int, bottomColor: Int) =
+        RoundedRectangleShader.performVQuadAndUniforms(
+            SkyHanniRenderPipeline.ROUNDED_RECT(),
+            x1 = left, y1 = top, x2 = right, y2 = bottom,
+            postVertexOps = listOf(
+                { color(topColor) },
+                { color(bottomColor) },
+                { color(bottomColor) },
+                { color(topColor) },
+            ),
+        )
+
     fun drawRoundedTexturedRect(left: Int, top: Int, right: Int, bottom: Int, texture: Identifier) =
         RoundedTextureShader.performVQuadAndUniforms(
             SkyHanniRenderPipeline.ROUNDED_TEXTURED_RECT(),
